requirements.txt
pytest==6.0.1
allure-pytest==2.8.18
urllib3== 1.25.10
requests==2.24.0
PySocks==1.7.1
Faker==4.1.1
pytz==2020.1
assertpy==1.1
deepdiff==5.0.2
dacite==1.5.1
python-dateutil==2.8.1
jsonschema==3.2.0
pydantic==1.8.2